requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
decorator==5.1.1
Django==3.2
django-extensions==3.1.5
django-haystack==3.1.1
djangorestframework==3.12.4
drf-haystack==1.8.11
elasticsearch==7.17.1
executing==0.8.3
ipdb==0.13.9
ipython==8.2.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
